+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 -   Ping command.
 -   Ip command.
+-   Ls command.
 
 ### Changed